requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pandas==2.1.3
numpy==1.26.2
opencv-python-headless==4.8.1.78
scikit-image==0.22.0
Pillow==10.1.0
gunicorn==21.2.0
